+++ lib/zoi/types/enum.ex
@@ -1,6 +1,6 @@
 defmodule Zoi.Types.Enum do
   @moduledoc false
-  use Zoi.Type.Def, fields: [:values, :enum_type]
+  use Zoi.Type.Def, fields: [:values, :enum_type, :coerce]
 
   def new(values, opts \\ []) when is_list(values) do
     type = verify_type(values)
@@ -50,8 +50,8 @@ defmodule Zoi.Types.Enum do
   end
 
   defimpl Zoi.Type do
-    def parse(%Zoi.Types.Enum{} = schema, input, _opts) do
-      case parse_enum(schema, input) do
+    def parse(%Zoi.Types.Enum{} = schema, input, opts) do
+      case parse_enum(schema, input, opts) do
         {:error, _reason} = error ->
           error
 
@@ -63,24 +63,16 @@ defmodule Zoi.Types.Enum do
       end
     end
 
-    defp parse_enum(schema, input) do
-      cond do
-        schema.enum_type in [:atom_binary, :binary] and is_binary(input) ->
-          compare_input(schema.values, input)
+    defp parse_enum(%Zoi.Types.Enum{values: values} = schema, input, opts) do
+      coerce = Keyword.get(opts, :coerce, schema.coerce)
 
-        schema.enum_type in [:atom_integer, :integer] and is_integer(input) ->
-          compare_input(schema.values, input)
-
-        schema.enum_type == :atom and is_atom(input) ->
-          compare_input(schema.values, input)
-
-        true ->
-          error(schema)
-      end
-    end
-
-    defp compare_input(values, input) do
-      Enum.find(values, fn {_key, value} -> input == value end)
+      Enum.find(values, fn {key, value} ->
+        if coerce do
+          input == value or input == key
+        else
+          input == value
+        end
+      end)
     end
 
     defp error(schema) do